// apps/api/src/modules/whatsapp/entities/whatsapp-appointment.entity.d.ts
export declare class WhatsappAppointment {
    id: string;
    clientId: string | null;
    phone: string;
    name: string | null;
    type: string;
    status: string;
    scheduledAt: Date;
    reminder24hSent: boolean;
    reminder2hSent: boolean;
    staffId: string | null;
    notes: string | null;
    metadata: object | null;
    feedbackRating: number | null;
    feedbackComment: string | null;
    createdBy: string | null;
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=whatsapp-appointment.entity.d.ts.map